+++ Yolov5_Onnx_Deploy.cpp
@@ -145,7 +145,7 @@ void Yolov5_Onnx_Deploy::post_image_process(std::vector<Ort::Value> &outputs, cv
 
     // NMS
     std::vector<int> indexes;
-    cv::dnn::NMSBoxes(boxes, confidences, 0.25, 0.45, indexes);
+    cv::dnn::NMSBoxes(boxes, confidences, (float)(0.25), (float)(0.45), indexes);
     for(size_t i = 0; i < indexes.size(); i++)
     {
         int idx = indexes[i];
