+++ Yolov5_Seg_Onnx.cpp
@@ -169,7 +169,7 @@ void Yolov5_Seg_Onnx::post_image_process(std::vector<Ort::Value> &outputs, cv::M
 
     // NMS
     std::vector<int> indexes;
-    cv::dnn::NMSBoxes(boxes, confidences, 0.25, 0.45, indexes);
+    cv::dnn::NMSBoxes(boxes, confidences, (float)(0.25), (float)(0.45), indexes);
     cv::Mat rgb_mask = cv::Mat::zeros(inputimage.size(), inputimage.type());
 
     for(size_t i = 0; i < indexes.size(); i++)
